% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFirstTwo.R
\name{addFirstTwo}
\alias{addFirstTwo}
\title{addFirstTwo}
\usage{
addFirstTwo(vec)
}
\arguments{
\item{vec}{a vector with numeric values}
}
\value{
Added first two elements of a vector
}
\description{
Function that returns the sum of the first two elements of a vector
}
\examples{
addFirstTwo(c(1,2))
addFirstTwo(c(3,4,5))

}
